tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.attributes.setAttribute.zero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.size.attributes.setAttribute.zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.attributes.setAttribute.zero</h1>
<p class="desc">Parsing of non-negative integers in setAttribute</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="50" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers in setAttribute");
_addTest(function(canvas, ctx) {

  canvas.setAttribute('width', '0');
  canvas.setAttribute('height', '0');
  _assertSame(canvas.width, 0, "canvas.width", "0");
  _assertSame(canvas.height, 0, "canvas.height", "0");
  _assertSame(window.getComputedStyle(canvas, null).getPropertyValue("width"), "0px", "window.getComputedStyle(canvas, null).getPropertyValue(\"width\")", "\"0px\"");
  _assertSame(canvas.getAttribute('width'), '0', "canvas.getAttribute('width')", "'0'");
  _assertSame(canvas.getAttribute('height'), '0', "canvas.getAttribute('height')", "'0'");

});
</script>
